op_alloy_protocol::batch
Module element

Span Batch Element

Structs

SpanBatchElement

A single batch element is similar to the SingleBatch type but does not contain the parent hash and epoch hash since spans do not contain this data for every block in the span.

Constants

MAX_SPAN_BATCH_ELEMENTS

MAX_SPAN_BATCH_ELEMENTS is the maximum number of blocks, transactions in total, or transaction per block allowed in a span batch.