op_alloy_protocol::batch
Module tx_data

Contains all the Span Batch Transaction Data types.

Re-exports

pub use wrapper::SpanBatchTransactionData;

pub use legacy::SpanBatchLegacyTransactionData;

pub use eip1559::SpanBatchEip1559TransactionData;

pub use eip2930::SpanBatchEip2930TransactionData;

Modules

eip1559 🔒

This module contains the eip1559 transaction data type for a span batch.

eip2930 🔒

This module contains the eip2930 transaction data type for a span batch.

legacy 🔒

This module contains the legacy transaction data type for a span batch.

wrapper 🔒

This module contains the top level span batch transaction data type.