op_alloy_protocol
Module fee

This module contains the L1 block fee calculation function.

Constants

NON_ZERO_BYTE_COST 🔒

ZERO_BYTE_COST 🔒

Cost per byte in calldata

Functions

calculate_l1_fee_scaled_ecotone 🔒

calculate_tx_l1_cost_bedrock

Calculate the gas cost of a transaction based on L1 block data posted on L2 post-bedrock.

calculate_tx_l1_cost_ecotone

Calculate the gas cost of a transaction based on L1 block data posted on L2, post-Ecotone.

calculate_tx_l1_cost_fjord

Calculate the gas cost of a transaction based on L1 block data posted on L2, post-Fjord.

calculate_tx_l1_cost_regolith

Calculate the gas cost of a transaction based on L1 block data posted on L2 post-regolith.

data_gas_bedrock

Calculate the data gas for posting the transaction on L1.

data_gas_fjord

Calculate the data gas for posting the transaction on L1.

data_gas_regolith

Calculate the data gas for posting the transaction on L1.

tx_estimated_size_fjord 🔒