op_alloy_protocol::test_utils
Struct CollectingLayer

pub struct CollectingLayer {
    pub storage: TraceStorage,
}

Available on crate feature test-utils only.

A subscriber layer that collects traces and their log levels.

Fields
storage: TraceStorage

The storage for the collected traces.

Implementations

impl CollectingLayer

pub const fn new(storage: TraceStorage) -> Self

Creates a new collecting layer with the specified storage.