op_alloy_protocol
Module utils

Utility methods used by protocol types.

Enums

OpBlockConversionError

An error encountered during OpBlock conversion.

Functions

cmp 🔒

flz_compress_len 🔒

Returns the length of the data after compression through FastLZ, based on

flz_match 🔒

hash 🔒

is_protected_v

Checks if the signature of the passed [TxEnvelope] is protected.

literals 🔒

read_tx_data

Reads transaction data from a reader.

set_next_hash 🔒

starts_with_2718_deposit

Returns if the given value is a deposit transaction.

to_system_config

Converts the OpBlock to a partial SystemConfig.

u24 🔒